requirements.txt
termcolor
pathlib
joblib
scikit-learn
bigquery
pyjanitor
math
tensorflow
dotenv